Improve risk logging format
Current risk logging format "INPLACERISK" is not really appropriate in
context of "migration" scenario; i.e. not "in-place upgrade" but upgrade
combined with migration to new hardware.

Use of the more general word "risk" solves that problem, but could lead
to filtering problems such as those described in previous commit.

Solution that preserves filtering abilities is to use generic wording
but specific formating.  Instead of:

    INPLACERISK: MEDIUM: foo bar baz

we now have:

    preupg.risk.MEDIUM: foo bar baz

which is easily filtered, readable and informative enough and consistent
with logging format.