fix XSL files to support new version of OpenSCAP
Openscap will support split of stderr and stdout outputs instead
of one merged output as was before. XSL files are modified
to correctly handle this change and in addition are still compatible
with the old one.

In addition
  - stdout is presented as "Additional info"
  - stderr is presented as "Logs"
for better description for reader. And tabs were replaced by spaces.

TODO: Old-style report could be modified yet for better
      visualisation of these two "signs" - like changed color
      or something like that.