add new preupg-diff tool
preupg-diff tool compares multiple Preupgrade Assistant XML reports from which one is new with unknown issues and the others have been already analyzed.

The tool helps to find new issues that emerged in the new report by filtering out those results that are the same in the new report and at least in one of the analyzed XMLs. The trimmed report is output in XML and HTML format and named result_diff.xml/html.