fix2 patch for data/preassessment/scripts.txt
In the patched data/preassessment/scripts.txt file on RHEL system
the "df" utility is used without "-P" option for POSIX format, that
guarantees each item of the output will be printed on the one line
only without wrapping. Otherwise the output can be wrapped
and generated data will be broken or incomplete in better case. This
commit fixes use of "df".

In addition, the "find" utility of version 4.2.27 (or lesser), does
not contain option "-executable", originally used in the scripts.txt
file. This is problematic on RHEL 5 system (or just systems which
uses too old find option). Instead of this, should be used:
   "-perm /111"
which is compatible with new version of the utility too.