RHEL5: fix python24 patch
Previously there were use of python3 "print" function rewrite
to use of python2 print function. But the rewrite changed the output,
as there was used stdout instead of stderr sometimes; sometimes
was produced string with two newlines ending instead of one newline
character. This is fixed now.

TODO: still are produced empty lines in output for stdout and stderr
      in result file