Fix logs and additional output in result file
Previously the report file result.xml contained whitespaces before
and after text of stdout and stderr (additional info and logs).
It is waste of place, so strip these characters.

In addition, "stdout" of modules is removed when it is empty, however
newly added "stderr" wasn't. Handle both of them in the same way.