remove the BEL characters when reloads XML file
The OpenSCAP put whole stdout/stderr into the result.xml directly
so we are not able to handle these outputs. In case that the BEL
character is produced by some utility (it happened already), parsing
of XML file fails.

NOTE: Some XML libraries handle it already, but not the one, used
      by the PA, mainly on older systems