Add module set init script
If module set contains an executable called 'init' (directly in its root
dir, "next" to properties.ini), it will be executed.  If the process
either exits with non-zero status or produces stderr, ModuleSetInitError is
raised.

This makes it possible to do module set specific steps before the module
set execution and possibly cause the process to stop, if the code
"decides" that it's not compatible.