Support older versions of the OpenSCAP
Old OpenSCAP doesn't support new complex report style that is used
by default by the Preupgrade Assistant tool. Enforce the simple
style of report (like with --old-report-style) when version of
installed OpenSCAP is lower then 1.2.7. In case of enforced fallback
to the simple style of report log message to inform the user.