Cleanup: remove "third_party" and dead code
Third party modules are not processed now. The code handling such
modules was messy and full of bugs, so it was impossible to process
such modules and tries about that was different even to original
guideline.

As a result all source code around third party modules is removed now
with additional found dead code to provide later cleaner and working
solution.