XCCDFCompose: add possibility to specify dst dir
Previously in case of building of XCCDF compose the dst directory was
strictly src directory with the "-results" suffix that was created
in same parent directory:
  /path/parent_dir/dirname
  /path/parent_dir/dirname-results

Now it is possible to specify dst path (including new dirname).
In case that dst path is not specified, behaviour is same as before.

TODO: handle situation when src and dst paths are same