Build XCCDF only from module.ini and group.ini files
Previously building of module set has been corrupted when directory
with the module contained more than one file with the ".ini"
extentions. Build of the module set shouldn't be affected by various
INI files, that are not related to the module set structure.

So the scan is now read only group.ini and module.ini files whose
names are reserved for PA purposes. In addition we expect only one
of those files in each directory, so the ModuleSetFormatError
exception is raised when both are detected in the same directory.

- removed unused import of ModuleSetUtils

Resolves #310