Check that openscap rpms are related to the system arch
OpenSCAP related rpms are built for 64b and 32b arch as well. The
issue occurs on 64b systems where even 32b verions of rpms could
be installed. In case that only 32b rpms are installed on 64b systems
the assessment of the system fails as all modules are skipped with
'notchecked' status.