Fix Web UI overall report risk level
Previously, for every uploaded report to the Web UI, the overall
"In-place Upgrade Risk" shown on the main page listing all preupg
runs was always "Slight". That was the default value. This issue
of using always the default value was that when parsing the pre-upgrade
result XML, the risks logged by the preupgrade-assistant modules were
not possible to find because of a change of an XML tag attribute. This
attribute has most probably changed due to 407cce7.

Resolves: BZ#1707913