Doc: update tutorial for custom modules development
Regarding the feedback, we are updating the current tutorial about
development of custom PA modules to reflect possibly the only use-case
to extend in-place upgrade from RHEL 6 to RHEL 7.

The main readme file for the tutorial is updated to provide a basic
overview about the PA modules and development. The existing cases
are extended by additional README files to provide info about each
module example.

Added several other templates for new modules:
    05_add_pre_upgrade_script
    06_add_post_upgrade_script
    07_module_in_python
    08_install_custom_rpms